refactor(web/engine): ruleBehavior.ts
